requirements.txt
opentelemetry-instrumentation-openai
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
psycopg2-binary